object-file-convert: Implement convert_object_file_{begin,step,end}
When converting trees, commits, and tags the objects they reference
need to be converted before the objects themselves can be converted.

Split convert_objet_file_convert into a couple of pieces that are
effectively an iterator over the oids that need to be converted.  This
allows the objects to be processed depth first when being converted
and it allows changing the logic to map oids.  In cases like "git
index-pack" none of the oids will be mapped in any of the existing
mapping tables so an in-memory table needs to be converted and
consulted, and this allows that.

Not having to update the existing object id mapping mechanisms
is particularly nice as it makes it easy to avoid having to
introduce new locks to syncrhonize the update of internal
mapping mechanisms.

This was inspired by a similar change by "brian m. carlson"
where he modified convert_object_file to return the
unmmaped oids.

Inspired-by: brian m. carlson <sandals@crustytoothpaste.net>
Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>