builtin/fast-import: compute compatibility hashs for imported objects
When the code is in dual hash mode for every object fast-import
creates compute the standard oid and it's compatibility mapping.  The
compatibility mapping is stored in struct pack_idx_entry so that it
can be used when an index is created.

For fast-import the code needs to be careful because when a new object
only refers to other newly created objects the compatibility mapping
for those new objects is not stored anywhere permanently.  So have the
code first look the the compatibility oid in the newly created
objects, and then look for the compatibilty oid in the standard
mapping tables.

As fast-import requires objects to be specified before the
objects that reference them nothing special needs to happen
to deal with out of order objects.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>